Seperated radPlotter into two macros: makeTree.C and radPlotter.C
As functionality of radPlotter grew, so did the clutter, so it has
been seperated into two macros. makeTree does what the 1st half of
the old radPlotter did, simply creating and saving a tree that holds
every fit parameter as a branch. The new radPlotter is now more aptly
names, as it handles all the plotting capabilities.

Some small changes such as converting large arrays to be vectors has
also occured, along with other readability improvements